fix the bar dia bug

+++ util.h
@@ -31,7 +31,7 @@ std::string Util::getProgressBar(std::string percent){
     int _size= 50;
     int  boundaries;
     try {
-        boundaries = int(stof(percent)/100.0f) *_size;
+        boundaries = stof(percent)/100.0f *_size;
     } catch (...){
     boundaries = 0;
     }
